updated sagemaker example
