--- examples/models/sagemaker/code/scikit_learn_iris.py
@@ -1,40 +1,0 @@
-#  Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
-#  
-#  Licensed under the Apache License, Version 2.0 (the "License").
-#  You may not use this file except in compliance with the License.
-#  A copy of the License is located at
-#  
-#      http://www.apache.org/licenses/LICENSE-2.0
-#  
-#  or in the "license" file accompanying this file. This file is distributed 
-#  on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either 
-#  express or implied. See the License for the specific language governing 
-#  permissions and limitations under the License.
-
-from sklearn.externals import joblib
-import numpy as np
-import os
-import logging
-
-
-logging.basicConfig(format='%(asctime)s %(levelname)s - %(name)s - %(message)s', level=logging.INFO)
-
-logging.getLogger('boto3').setLevel(logging.INFO)
-logging.getLogger('s3transfer').setLevel(logging.INFO)
-logging.getLogger('botocore').setLevel(logging.WARN)
-
-logger = logging.getLogger(__name__)
-logger.setLevel(logging.DEBUG)
-
-def model_fn(model_dir):
-    print("LOOOK--------------> ",model_dir)
-    clf = joblib.load(os.path.join(model_dir, "model.joblib"))
-    return clf
-
-def predict_fn(data, model):
-    print("Predict")
-    if len(data.shape) == 1:
-        data = np.reshape(data,(1,data.shape[0]))
-    return model.predict(data)
-
-
