--- examples/models/sagemaker/requirements.txt
@@ -1,3 +1,0 @@
-sagemaker
-sagemaker-containers
-pandas
